[WIP] Reduce CML interface to be minimal
In order to prepare Trust|me to be evaluated according to CC
we need to reduce all interfaces to the absolute minimum.

This commits reduces the interfaces that are implmented via the
control and service sockets. To avoid having to maintain a hard
fork, a new preprocesser flag, "CC_Mode" is being introduced.
This flag will effectively filter commands that the cmld services.

In order for clients to be able to implement the reduced interfaces
we supply the specification in form of adapted .proto files.

TODOs:
- implement control command filtering in daemon/control.c
- reflect reduced command set in daemon/control.proto
- implement reduced service interface in cmld
- properly comment all .proto files so that developer documentation
  can be built from them automatically